Add SConstructs and SConscripts to build WALi.
The SConscript to build the WALi library depends on following
SCons variables.

Import('WaliDir')
Import('LibInstallDir')

To use the SConscript for an external project, define these two variables. 
LibInstallDir tells teh SConscript where to install the library.
WaliDir is the root of the WALi hierarchy.

For those with more SCons experience than me, some tidying is appreciated :)

Why -- /s/std/bin/libtool is a piece of ****!




git-svn-id: file:///p/wpis/people/students/kidd/svnwali/trunk@308 dfa36f5b-b2fa-0310-9605-e5aca7e64462