Made FWPDS return a WFA with lazy weights that are evaluated on demand.
Added class LazyTrans that keeps a pointer to InterGraph and evaluates weight
on the transition on demand. This required some changes in Trans to make sure
that weight is always accessed through weight()

Added a filter(n) method in WFA to remove transitions such that the resultant WFA is
as if an intersection with (n \gamma^*) was performed.

To use lazy FWPDS:
poststar(input,output);
output.filter(node);
output.path_summary();
return output.getState(init_state)->stateWeight();



git-svn-id: file:///p/wpis/people/students/kidd/svnwali/trunk@346 dfa36f5b-b2fa-0310-9605-e5aca7e64462