Create class GenKeySource that tags a wali::Key with a generation count.
Added 'size_t generation' to wfa::WFA that defaults to the value 0.

WPDS::setupOutput (called by WPDS::setup*FixPoint methods) increments the
generation of the output WFA by one.

WPDS::gen_state(Key k1,Key k2) creates a generational key source using the
generation of the output automaton.

EWPDS::poststar_handle_eps_trans first checks if a key is a KeyPairSource and if
that fails if it is a GenKeySource containing a KeyPairSource.


git-svn-id: file:///p/wpis/people/students/kidd/svnwali/trunk@350 dfa36f5b-b2fa-0310-9605-e5aca7e64462