Create an interface class IMergeFn.   ** This allows one to distinguish between a default MergeFn      or a user MergeFn. (It would be nice if the class      MergeFn could be declared final as in Java.)
EWPDS will create a new MergeFn if the same push rule is added
twice. The new MergeFn has a weight that is equal to 
the weight of the push rule after WPDS has handled it. 
(I.e., let WPDS handle wrapping and semiring combines.)

If the same push rule is added twice and the user supplied
merge functions, then an error is printed and WALi "throws"
the push rule. The above exception was added b/c Empire
generates such WPDSs (argh).


git-svn-id: file:///p/wpis/people/students/kidd/svnwali/trunk@442 dfa36f5b-b2fa-0310-9605-e5aca7e64462