Refactor the parse code so that all Handlers accept a reference to an IUserHandler.
QueryHandler delegates to WfaHandler and WpdsHandler.

(Wfa|Wpds|EWpds)Handler either take action, or if the element
tag is not recognized, the IUserHandler is queried to see if it
handles the element. If so, the delegation proceeds.

UserFactoryHandler is a class for backwards compatability with
parsing using WeightFactories and MergeFnFactories.
For this to work properly, it is important that the method
DefaultHandler::characters delegates to the fUserHandler inside
of (Wfa|Wpds|EWpds)Handler. This is brittle and hopefully we
can come up with a better solution. Note that it works b/c
a handler defined by WALi does not use characters, and should
delegate to the IUserHandler so that it can function propertly.

Ben & Cindy -- This should cover the parsing that we discussed.
We should finish the XSD files and add them to the repos.

Cindy -- the problem with witnesses should also be fixed. 
We can work this out when you are ready.


git-svn-id: file:///p/wpis/people/students/kidd/svnwali/trunk@465 dfa36f5b-b2fa-0310-9605-e5aca7e64462