Added the ability to force a 32-bit or 64-bit build from the command line running SCons & bug fix
Just add 'arch=XXX' where XXX is either 32, 64, or any of several common names
for each architecture. (See the ThirtyTwoBitAliases and SixtyFourBitAliases
variables in SConstruct.)

Also fixed a bug where 'm32' instead of '-m32' was passed for a 32-bit build
on a 64-bit system (thus causing the linker to complain that it couldn't find
'm32').


git-svn-id: file:///p/wpis/people/students/kidd/svnwali/branches/burtona@814 dfa36f5b-b2fa-0310-9605-e5aca7e64462