This should correctly set the 'modified()' flag in a Trans and avoid extra work.
Here is an email sent to Nick which he then said sounds correct:

Tom and I were looking through the poststar code trying to figure out
the mystery about the weight-based priority queue with shortest-path
weights being slower than the key-order worklists, and Tom found
something which we think could be a bug. (It is a "do too much work"
bug, not a "wrong answer" bug.) It shouldn't be causing the mystery with
the worklist priorities though. We want to know what you think.

This concerns the 'modified()' function inside the Trans class. Here's
what the comment on its declaration in ITrans says:

    /*!
     * @brief return true if the transition has been modified
     *
     * A ITrans is considered modified if when its weight changes. This
     * includes the initial creation of a ITrans object. This follows
     * from the fact that all ITrans can be considered (abstractly)
     * created with a weight of ZERO
     *
     * @return true if this transition has been modified
     */
    virtual bool modified() const throw() = 0;

(Incidentally, this comment is missing the answer to a huge question
which is pretty much necessary for using it: "An ITrans is considered
modified if its weight changes..." since when?) The code which uses it
seems to be assuming this behavior. There is one use in WPDS's
'update()' function:

    void WPDS::update(Key from, Key stack, Key to,
                      sem_elem_t se, Config * cfg)
    {
      wfa::ITrans*t = currentOutputWFA->insert(new Trans(from,stack,to,se));
      t->setConfig(cfg);
      if (t->modified()) {
        //t->print(std::cout << "Adding transition: ") << "\n";
        worklist->put( t );
      }
    }


and one inside 'poststar_handle_trans()'.


However, looking at where the Trans's status is set, inside
'combineTrans()', we think it's doing something different:

    void Trans::combineTrans( ITrans* tp )
    {
      sem_elem_t wnew = tp->weight();

      // delta returns ( wnew + se, wnew - se )
      // Use w->delta(se) b/c we want the returned diff
      // to be what is in the new weight (wnew) and not
      // in the existing weight (se)
      std::pair< sem_elem_t , sem_elem_t > p = wnew->delta( se );

      // This's weight is w+se
      se = p.first;

      // Delta is combined with the new delta.
      delta = delta->combine( p.second );

      // Set status
      status = ( delta->equal(delta->zero()) ) ? SAME : MODIFIED;
    }


We think that 'status' should be set according to whether 'delta' has
*changed*, not whether 'delta' is 0, so something like this:

      sem_elem_t old_delta = delta;
      delta = delta->combine(p.second);
      status = ( old_delta->equal(delta) ) ? SAME : MODIFIED;

Our mental model of how the default implementation of 'delta' works
(which shortest path uses currently -- perhaps this should change?) is
that 'delta' is always either equal to the transition's weight ('se') or
its semiring zero, and that 'delta != 0' should be true exactly when the
transition is in the worklist (or is being processed).

Thinking about it as I type this, the 'update()' function think
shouldn't suffer because the worklist's 'put' function will see that the
transition is already marked and not do anything. (Though with a
"proper" implementation, 'put()' would have to call 'decreaseKey()' in
the case where the element isn't marked, and getting 'modified()' right
would save the call to 'decreaseKey()'.) But the
'poststar_handle_trans()' function seems like it might wind up doing
rather more work than is necessary.




git-svn-id: file:///p/wpis/people/students/kidd/svnwali/trunk@1572 dfa36f5b-b2fa-0310-9605-e5aca7e64462