Fix the test fixture. How did this ever work?!
What happened was that one of the fixtures got a bunch of keys right away and stored them in static members. But then the Wali prerequsites tests ran, and as part of those tests, they cleared the key space. This essentially orphans those keys which were in static members before. (Again: how did this ever work?!)

git-svn-id: file:///p/wpis/people/students/kidd/svnwali/trunk@1667 dfa36f5b-b2fa-0310-9605-e5aca7e64462