Total restructure of the PriorityWorklist class-NWA
Instead of having a virtual 'compareTo()' function which gets overridden, PriorityWorklist is now a template. The template argument is the comparator.

KeyOrderWorklist and TotalOrderWorklist needed to change to match. Now they define the comparison functor and inherit from PriorityWorklist<that comparator>. This leaves those classes essentially empty.

The reason behind this is that, unless I did something wrong (and I might have -- one possibility occurred to me as I was writing this, and I may still back out this change if it is borne out), the Boost heap library requires your comparators to have default constructors. That wouldn't work in the old design, as the comparator has to have a pointer (or reference) back to the PriorityWorklist so it can call its compareTo function.

git-svn-id: file:///p/wpis/people/students/kidd/svnwali/trunk@1671 dfa36f5b-b2fa-0310-9605-e5aca7e64462