Made the Witness class cache zero and one values instead of creating a new object each time zero()/one() is called.
In almost all cases this will result in a nice improvement in the amount of
memory (and also speed) used. However, it can be a regression if the weights
the witnesses are wrapping don't cache their own zeros and ones, or if many
many many different kinds of weights are used.



git-svn-id: file:///p/wpis/people/students/kidd/svnwali/trunk@1926 dfa36f5b-b2fa-0310-9605-e5aca7e64462