Add a new "longest path" weight domain (it saturates)
No, it's not yet April 1. I need a test for epsilon close that 
illustrates that epsilon loops caused the wrong weight to be returned. 
For a simple automaton

        a
   A ------> B
     <------
        b

where both transitions are epsilon and the a/b are weights, I needed a 
domain where (a != a + a*b*a). Relational weights will do it, but dealing 
with them is really heavyweight, so I introduced this "longest distance" 
weight. :-)



git-svn-id: file:///p/wpis/people/students/kidd/svnwali/trunk@1986 dfa36f5b-b2fa-0310-9605-e5aca7e64462