Fix documentation bug re. weight generation for epsilon transitions
The code does not do anything special for epsilon transitions when converting an NWA
to WPDS, just calls getWeight() like for any other symbol. (It was even more wrong
before, because it explicitly talked about what the weight would be for epsilons on
calls and returns, which can't happen.)

This may not be the final word on this error, but it will suffice for now.