Introduce new optional SemElem virtual functions hash() and containerLessThan()
These are used by SemElemSet; which one is used depends on the backing set type,
and will be used later. I suggest not overriding these functions if they cannot
operate quickly.

(containerLessThan() imposes an arbitrary strict weak order on semiring elements;
different pointers to the same element need to behave consistently and it needs to
be consistent with equals(), but aside from those requirements that's all that
needs to be met. In particular, semelem implementers needn't make it relate to the
semantic meaning, and users should not assume that it does.)

Both functions abort() in their default implementation.