Implemented Push based RegExp evaluation scheme
When the macro PUSH_EVAL is defined, a mixed push/pull based regexp evaluation
is used. This saves further on evaluation of the dag.
Nodes above updatable edges are set as dirty, and only dirty nodes are
considered during evaluation. The existing pull based machinery is still
maintained and used to determine candidates for evaluation.