Changed bdd variable order to TENSOR_MIN_AFFINITY
bdd tensor is extremely slow under the earlier order of MATCHED_PAREN. That order was needed for the nwa based detensor. It does not make sense of normal detensor anyway, so going back to TENSOR_MIN_AFFINITY, which is known to be the fastest of all available orders at this point.