Change STATIC_MEM in IntraGraph to INTRAGRAPH_SHARED_MEM. It no longer uses static variables.
IntraGraph earlier used static variables so that different IntraGraph objects could share memory buffers. Added a struct IntraGraph::SharedMemBuf to encapsulate these variables. An object of this type can be passed to IntraGraph constructor, and it will then use the object if the macro INTRAGRAPH_SHARED_MEM is defined. This behaviour is the same as before except for the renaming of the macro. Changes
(1) Renamed macro STATIC_MEM to INTRAGRAPH_SHARED_MEM
(2) Eliminated some static variables in IntraGraph.