ProgramBddContext::Assume bug fix for different voc sizes
Must quantify out the regSize from bdd before conjucting the two operands.
Constants take the size of the largest variable in the vocabulary. This would
lead to Assume returning false. Now, the semantics are the equate bits equal to
the smaller of the two operands to Assume, and leave the higher bits
unconstrained.