Partial changes to incorporate accepting weights to WFAs
WFA states are no longer (well, sort of) binary accepting or not accepting, but now have an acceptingWeight(). When you add a new accepting state, it adds that state with accepting weight of 1. The semantics should be that a state is "accepting" iff its accepting weight is nonzero.

WFA determinize weight gen now has a function that will return the weight to be used on accepting states.