Two non-failing tests & safety fix for zero weights
I thought I was getting incorrect results because zero-weight paths were being considered accepting, so I made a change to omit states reachable via zero-weight paths from eclose and wrote a couple tests. But I think the problem turned out to be elsewhere, so these changes are not so important.