Add DeterminizeWeightGen::getOne function
The WFA determinize function needs an example weight to label nodes of the resulting automaton. It was using a weight from the input, but this is often going to be incorrect, as the weights can be of different types. Instead, it will now call this function. (It does not do so yet, but will next commit.)