Remove assertion which is too strong
I'm actually rather surprised this hasn't fired before now. Basically it was depending a bit on an order of stuff in an unordered_map a bit. The assertion checked that "adding" something to a SemElemSet increased the size by 1, but that isn't necessarily the case if the new element subsumes existing ones.