Prep for AWPDS: use WFA weights instead of PDS weights for WFA stuff
This is so that there can be different types of weights on the PDS and WFA.

This commit is a fixed version of what was originally a small part of commit 62bfe12 on what is now the defunct-branch-awpds (and then 2c9e757 on defunct-branch-awpds-rebase-and-fix) tag. The fix is f959d on the defunct-branch-awpds-rebase-and-fix branch.