WFA::path_summary: fix bug relating to extend order
Suppose there is a WFA

           a/w_1       b/w_2
   -> (a) ------> (b) ------> ((c))

Then there are two possibilities for the weight that should label (a). If the 'query' member is set to 'INORDER', it should be w_1*w_2. If the 'query' member is set to 'REVERSE', it should be w_2*w_1. (Or perhaps that's backwards.)

Previously, the WPDS-based path_summaries did not pay attention to the query direction, which of course was wrong.