BoolMatrix becomes a template again
The definition is moved to Matrix_template.hpp, with a typedef in Matrix.hpp.
Matrix_template.hxx has the implementation, but it is NOT automatically included.

An instantiation for 'bool' is provided in Matrix.cpp/hpp. Later, other
instantiations will be provided. You can use this in the same ways as current:
build Matrix.cpp and #include Matrix.hpp.

Also rename test files.