Reach the goal for matrix shortest path
Modifies ShortWitnessVisitor so that it can take a ranker;
when it hits a combine node, it now picks a child with the
lowest rank. There is now a MatrixRanker that can extract
out specific elements from a matrix, so you can request only
paths that result in an error state or whatever.