Revert "Change SemElem::delta()'s return value"
This reverts commit 55b531abea242d13727b22ca386389d8d6fda7a2.

The motivation behind the commit was to change the implementation of SemElem::delta from (i) <this + se, this + se = se? 0 : this + se>
to (ii) <this + se, this + se = se? 0 : this>. The motivation behind this change is not clear (Evan guesses that it might be related to providing support to Junghee's AWPDS).

While approach (ii) is sound for distributive semirings, it can be unsound for non-distributive semirings. This changes introduces unsoundness in analysis with non-distributive domains such as KS and domains with infinite ascending chains like polyhedra which break the distributivity by performing non-monotonic widening operation instead of combine in the presence of loops.