Fix bug in WFA::removeEpsilons -- now adds correct final states
If you have an epsilon path
   A ----> B ----> C ----> D
where D is final, then A should be final in the transformed WFA.

I tried to do this, but did it badly, and sometimes things would be
dropped. For example, it would work if it removed those edges in
the order C->D then B->C then A->B, but not other orders.