Explicitly cast a Boost optional to bool where needed
This used to happen implicitly wherever needed, but at some point
Boost changed the conversion operator to be explicit.  That means it
will happen automatically only in some contexts, such as "if"
conditions.  It needs more explicit urging elsewhere.

This resolves one warning when running "scons test" under Fedora 27,
using gcc 7.3.1.  That warning is configured to be treated as an
error, so this test was actually failing on the aforementioned
platform.